refactor(Formatter): 패키지 위지 변경(dto -> util)
