+++ src/main/java/util/Formatter.java
@@ -1,4 +1,4 @@
-package dto;
+package util;
 
 import java.time.format.DateTimeFormatter;
 import java.util.Locale;
@@ -8,5 +8,5 @@ public class Formatter {
     public static final DateTimeFormatter TIME_FORMATTER = DateTimeFormatter.ofPattern("HH:mm", Locale.KOREAN);
     public static final DateTimeFormatter DATE_FORMATTER = DateTimeFormatter.ofPattern("MM월 dd일 E요일", Locale.KOREAN);
     public static final DateTimeFormatter DATETIME_FORMATTER = DateTimeFormatter.ofPattern("MM월 dd일 E요일 HH:mm",
-        Locale.KOREAN);
+            Locale.KOREAN);
 }
